Fix the link server settings page
Currently the library link server settings page sets the server to the name when applying settings.  However, when it's used by the linkGenerator it needs to be the URL.  In BBB, users have their settings as the URL, so this needs to be updated.

This may affect some users who updated their settings, so if they aren't getting a "My Institution" link, would need to refresh their link server.

Do we need to handle the case that the incoming link_server is just the name? in that case it means fetching the link servers again and extracting the corresponding URL.