Fix settings error on abstract page
When navigating to abstract page, if the user is not logged in, the 'useGetUserSettings' call will return an undefined value.

I missed this was the hook used here in an earlier commit.  I should have switched this to use `useSettings` since this hook guarantees the `settings` will return an object.