Stop replacing url on notification close
The toast notification for login was doing a shallow replace to update
the URL.  But if the user makes a route change prior to the closing of
the toast, the original reset method is called using the old route, so
we redirect back a page.

To fix this I just took off the URL changes, since it's not actually
necessary.