Fix unnecessary requests issue
Seems that sometimes the app can get into a request retry loop and it can cause cascading requests.

Part of the issue is that in the case of 401s a separate (from RC) handler performs it's own re-request.

This disables retrying for now until I can figure out the right mix.