Update the API request timeouts for SSR
Changed the serverside API call timeout to 1.5 seconds (down from 30
seconds, 5% of the clientside calls)
Clientside should refetch it and show a loading state if the query fails
to load in time.

This timeout value will need to be tested, but the percieved speed is
definitely improved with a short server-side timeout.

For URL array params switch from `repeat` to `comma` and update tests
that broke.