Change default sort for library search view
When users opt to view a library as a search query, it defaults to sorting by relevance. Users have sent feedback that this is confusing, asking that we instead sort by date descending. This mimics the sort used in the library view.

Feedback link: https://v2.hiverhq.com/permalinks/pvt/1dc4a1a8